+++ src/main/java/vn/BE_SWP302/domain/response/DashboardResponse.java
@@ -3,8 +3,6 @@ package vn.BE_SWP302.domain.response;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 public class DashboardResponse {
@@ -12,8 +10,8 @@ public class DashboardResponse {
     private Long totalDoctors;
     private Long totalBookings;
     private Long todayBookings;
-    private BigDecimal todayRevenue;
-    private BigDecimal monthRevenue;
+    private Double todayRevenue;
+    private Double monthRevenue;
     private Long totalServices;
     private Long totalFeedback;
     private Long ongoingTreatmentCount;
