+++ src/main/java/vn/BE_SWP302/domain/response/InvoiceResponse.java
@@ -1,6 +1,5 @@
 package vn.BE_SWP302.domain.response;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import lombok.Getter;
@@ -11,6 +10,6 @@ import lombok.Setter;
 public class InvoiceResponse {
     private Long invoiceId;
     private LocalDateTime issuedDate;
-    private BigDecimal totalAmount;
+    private Double totalAmount;
     private String status;
 }
